requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-database-url==1.0.3
django-debug-toolbar==4.3.0
django-filter==24.2
django-safedelete==1.3.3
django-shortuuidfield==0.1.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==25.0.0
filelock==3.13.3
identify==2.5.35
inflection==0.5.1
nodeenv==1.8.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.7.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
shortuuid==1.0.13
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
virtualenv==20.25.1
